Refactor & move utilities. So that UDP can use them too.
